move UI manager to separate file
